requirements.txt
# pip3 install -r requirements.txt

requests
bs4
Pillow
pandas
